Import dragonboard audio library
The audio directory from the dragonboard git[1] is imported
as-is from the following commit:
    bc02e153097048c72e86037920a19a286fb012ca
    bc02e15 Snap for 6191568 from 7f47f75eaf9d475bbcdd78343e9a40cf6d986874 to rvc-release

For some reasons merely adding the following to midas.mk:
  PRODUCT_PACKAGES +=  audio.primary.exynos4
wasn't enough to get an audio.primary.exynos4.so library in the
out/target/product/i9300 directory.

Importing this audio library here fixes that.

In addition, if we look at the audio library in audio being used in
Replicant 6 for the GT-I9300[2], in audio/ we have calls to several
functions related to the RIL, like ril_open, ril_close,
ril_register_set_wb_amr_callback, ril_set_call_audio_path, and so on.

Because of that we will most probably need to implement something
similar to make audio work in calls, so it's probably best to fork
that library anyway, at least for the GT-I9300.

References:
-----------
[1]https://android.googlesource.com/device/linaro/dragonboard
[2]https://git.replicant.us/replicant/device_samsung_i9300.git

Thanks to Joonas Kylmälä who identified that the dragonboard
audio library worked out of the box on Android 11 for the
Galaxy SIII 4G (GT-I9305).

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>