midas.mk: mark sections topics to regroup things and enable more comments
This regroup statements in sections that are now indicated like that:
    ############
    # Graphics #
    ############

    PRODUCT_PACKAGE += [...]

Regrouping things in sections enable to avoid duplication like
we currently have with graphics for instance.

In addition, before the topic sections were using simple comments
like that:
    # Graphics
    PRODUCT_PACKAGE += [...]

And so it wasn't possible for an item to both add comments
detailing why it's done in this way and an indication of the
section it's in (like Graphics).

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>