audio: add extremely basic routing
The configuration is derived from the default and speaker on
configurations from the configs/tiny_hw.xml file in the
replicant-6.0 branch of the device_samsung_i9300[1].

This should be sufficient for testing the audio as it routes
it to the speaker and never changes route.

This will be improved in subsequent commits as only having the
ability to use the speaker and leaving them on all the time
doesn't really make the device usable.

The audio_hw.c part was inspired from the audio_hw.c and audio_hw.h
files in the device_amlogic_yukawa repository[2] at the following
commit:
    7252a5cd25ced5badee99e7377fb678394c943a8
    7252a5c Snap for 6453963 from b70ebf49d8bc1f162601b8cb627edfc7a689b67a to rvc-release

References:
-----------
[1]https://git.replicant.us/replicant/device_samsung_i9300.git
[2]https://android.googlesource.com/device/amlogic/yukawa

Thanks to Joonas Kylmälä who found that some boards were using
a mixer_paths.xml configuration file. With that information I
was able to find the libaudioroute library and an example xml
file.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>